requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.2.0
numpy>=1.25.0
Pillow>=10.0.0
matplotlib>=3.8.0
segmentation-models-pytorch>=0.3.1
torchmetrics>=1.0.0
